Extract HttpRequestHandler from HttpServer

+++ src/main/java/no/kristiania/pgr203/HttpHeaders.java
@@ -35,7 +35,7 @@ class HttpHeaders {
         addLine(fieldName + ": " + fieldValue);
     }
 
-    void setContentLength(int length) {
+    void setContentLength(long length) {
         add("Content-Length", String.valueOf(length));
     }
